装备随机词条

+++ Web/Data/Config/ItemConfig.cs
@@ -21,18 +21,20 @@ public class EquipmentConfig
     public int Quality { get; set; }
     public int Part { get; set; }
 
-    // 当前 Equipment.json 里这些字段是 [min,max] 数组或 null
-    public int[]? AttackRange { get; set; }
-    public int[]? HPRange { get; set; }
-    public int[]? DefenseRange { get; set; }
-    public int[]? Critical { get; set; }
-    public int[]? AttackSpeed { get; set; }
-    public int[]? CriticalDamage { get; set; }
-    public int[]? UpperLimbRange { get; set; }
-    public int[]? LowerLimbRange { get; set; }
-    public int[]? CoreRange { get; set; }
-    public int[]? HeartLungsRange { get; set; }
-public int Random { get; set; }
+    // NOTE: 最新 Equipment.json 中存在小数（如 0.0 / 5.0），因此这里必须用 double[]
+    // 读取后在生成装备实例时再按需要转 int。
+    public double[]? AttackRange { get; set; }
+    public double[]? HPRange { get; set; }
+    public double[]? DefenseRange { get; set; }
+    public double[]? Critical { get; set; }
+    public double[]? AttackSpeed { get; set; }
+    public double[]? CriticalDamage { get; set; }
+    public double[]? UpperLimbRange { get; set; }
+    public double[]? LowerLimbRange { get; set; }
+    public double[]? CoreRange { get; set; }
+    public double[]? HeartLungsRange { get; set; }
+
+    public int Random { get; set; }
 }
 public class EquipmentRandomConfig
 {
